+++ __tests__/setup/createUserId.js
@@ -1,0 +1,3 @@
+export default function createUserId() {
+  return `dl_${ Math.random().toString(36).substr(2, 5) }`;
+}