+++ __tests__/setup/getEchoActivity.js
@@ -1,0 +1,9 @@
+export default function getEchoActivity(activity) {
+  const {
+    channelData: {
+      originalActivity
+    } = {}
+  } = activity;
+
+  return originalActivity;
+}