+++ __tests__/setup/waitForBotToRespond.js
@@ -1,0 +1,5 @@
+import waitForObservable from './waitForObservable';
+
+export default async function waitForBotToRespond(directLine, predicate) {
+  return await waitForObservable(directLine.activity$, activity => predicate(activity));
+}