+++ __tests__/setup/waitForConnected.js
@@ -1,0 +1,11 @@
+import { ConnectionStatus } from '../../src/directLine';
+
+import waitForObservable from './waitForObservable';
+
+export default async function waitForConnected(directLine) {
+  const subscription = directLine.activity$.subscribe();
+
+  await waitForObservable(directLine.connectionStatus$, ConnectionStatus.Online);
+
+  return subscription.unsubscribe.bind(subscription);
+}