+++ webpack-development.config.js
@@ -1,0 +1,6 @@
+const webpackConfig = require('./webpack.config');
+
+module.exports = {
+  ...webpackConfig,
+  mode: 'development'
+};