+++ webpack.config.js
@@ -1,17 +1,17 @@
-const { resolve } = require('path');
 const { DefinePlugin } = require('webpack');
+const { join } = require('path');
 const Visualizer = require('webpack-visualizer-plugin');
 
 module.exports = {
   entry: {
-    'directLine': './built/directLine.js'
+    directline: './lib/directLine.js'
   },
   mode: 'production',
   output: {
     filename: '[name].js',
     library: 'DirectLine',
     libraryTarget: 'umd',
-    path: __dirname
+    path: join(__dirname, 'dist')
   },
   plugins: [
     new DefinePlugin({